Finished with the createUser use case for now.
