+++ server/modules/user/UserObject.type.ts
@@ -1,0 +1,8 @@
+type UserObject = {
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+};
+
+export default UserObject;
